Move utils to own folder
This makes it clear which files contain the public API and which don't
