+++ lib/__tests__/getLineAnchorForOffset.js
@@ -1,4 +1,4 @@
-var getLineAnchorForOffset = require('../getLineAnchorForOffset');
+var getLineAnchorForOffset = require('../utils/getLineAnchorForOffset');
 
 describe('getLineAnchorForOffset', function() {
