+++ lib/handleKeyCommand.js
@@ -1,4 +1,4 @@
-var removeIndent = require('./removeIndent');
+var removeIndent = require('./utils/removeIndent');
 
 /**
  * Handle key command for code blocks
