+++ lib/handleReturn.js
@@ -1,5 +1,5 @@
 var Draft = require('draft-js');
-var insertNewLine = require('./insertNewLine');
+var insertNewLine = require('./utils/insertNewLine');
 
 /**
  *  We split code blocks only if user pressed Cmd+Enter
