+++ lib/handleTab.js
@@ -1,5 +1,5 @@
 var Draft = require('draft-js');
-var getIndentation = require('./getIndentation');
+var getIndentation = require('./utils/getIndentation');
 
 // TODO: tab should complete indentation instead of just inserting one
