+++ lib/utils/getLineAnchorForOffset.js
