+++ lib/utils/getLines.js
