+++ lib/utils/getNewLine.js
