+++ lib/utils/insertNewLine.js
