+++ lib/utils/removeIndent.js
